Refactor : 트레이닝 등록 페이지 컴포넌트 분리 및 파일 구조 정리
