+++ src/components/trainee/training/DoughnutChart.vue
