+++ src/components/trainee/training/RoutineChat.vue
