+++ src/components/trainee/training/RoutineResultModal.vue
