+++ src/components/trainee/training/RoutineSection.vue
