+++ src/components/trainee/training/RoutineSubmitBar.vue
