+++ src/components/trainee/training/RoutineVideo.vue
